Update Tuning.py
fixed it so now it more accurately reflects our rocket where we are controlling the force exerted (acceleration) instead of directly changing the position. This fricked up all the calculation when it doesn't have an overshoot tho. Like it just won't calculate the steady state or overshoot as less than 1 even when it never reaches the target